mqtt_webcam_assignment/requirements.txt
paho-mqtt==2.2.1
opencv-python==4.8.0.74
numpy==1.24.3
pillow==10.0.0
